split modal and form components from params
